chore(tests): move review action test in its own file
