+++ mergify_engine/tests/functional/actions/test_review.py
@@ -1,0 +1,54 @@
+# -*- encoding: utf-8 -*-
+#
+# Copyright © 2020 Mergify SAS
+#
+# Licensed under the Apache License, Version 2.0 (the "License"); you may
+# not use this file except in compliance with the License. You may obtain
+# a copy of the License at
+#
+#      http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
+# WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
+# License for the specific language governing permissions and limitations
+# under the License.
+import yaml
+
+from mergify_engine.tests.functional import base
+
+
+class TestReviewAction(base.FunctionalTestBase):
+    def test_review(self):
+        rules = {
+            "pull_request_rules": [
+                {
+                    "name": "approve",
+                    "conditions": [f"base={self.master_branch_name}"],
+                    "actions": {"review": {"type": "APPROVE"}},
+                },
+                {
+                    "name": "requested",
+                    "conditions": [
+                        f"base={self.master_branch_name}",
+                        "#approved-reviews-by>=1",
+                    ],
+                    "actions": {
+                        "review": {"message": "WTF?", "type": "REQUEST_CHANGES"}
+                    },
+                },
+            ]
+        }
+
+        self.setup_repo(yaml.dump(rules))
+
+        p, _ = self.create_pr()
+
+        self.wait_for("pull_request_review", {}),
+
+        p.update()
+        comments = list(p.get_reviews())
+        self.assertEqual(2, len(comments))
+        self.assertEqual("APPROVED", comments[-2].state)
+        self.assertEqual("CHANGES_REQUESTED", comments[-1].state)
+        self.assertEqual("WTF?", comments[-1].body)
